Develop (#187)
* IFIX-714: code Change time zone.

* Change in time zone

* time zone issue.

* timezone change.

* IFIX-686

* IFIX-714: code Change time zone.

* Ifix-717

* IFIX-716: change by adding localization message for update.

* query fixes for demand and expense

* added fix for pdf demand and expnce values

* grammer sql exceptin resolved.

* IFIX-686

* IFIX-714: code changes for time period in demand.

* IFIX-720

* Change in condition to call the demand search

* IFIX-691-Added additional property details in event list model

* IFIX-720

* IFIX-720

* IFIX-720

* IFIX-714: UAT change

* added NICSms changes.

* added value for entityId in application.properties

* added log

* Update MgramsevaAdapterDemandConsumer.java

* added constant.

* added update demand logic in adaptor.

* change in common name param

* Added new param

* Change in config path

* IFIX-718

* modified code for template id appending.

* IFIX-718: Code changes related to pdf cumulative value.

* commented request method.

* IFIX-686

* IFIX-691

* IFIX-718: Code changes related to pending collection.

* Code changes related to demand not exist connections.

* IFIX-718: Code changes for time..

* IFIX-686

* IFIX-686

* IFIX-686

* ifix 703 dashboard changes.

* added if condition for checking length

* IFIX-713: code change in amount format.

* IFIX-713: code change in amount format.

* IFIX-718
update demand issue fixed

* IFIX-713: code change in amount format.

* IFIX-713: code change in amount format.

* Delete CronJobFNE.yaml

* Delete CronJobMonthly.yaml

* Delete failedbulkdemand.yaml

* change in connection no query

* nic sms fix

nic sms fix

* SMS related change in echallan schedular.

* Added extra condition to generate new demand.

* Nic SMS fixes

Nic SMS fixes

* Ng punjab mgramseva develop v5 (#79)

* added logs for test adaptor request.

* added checks for paid amount is zero for create payment topic adaptor.

* adapter changes fixes for bill payment.

* removed else condition

* adapter changes for duplicate bill generation.

* added db config for adpator.

* adpator fixes.

* removed logs statement.

* added cancel payment topic for listen.

* IFIX-741
IFIX-738
IFIX-740
IFIX-735

* fix for monthly summary msg

* fix for monthly summary msg (#87)

* added cancel payment logic.

* ifix-281 plane search for ws and challan service.

* Create txt

* nic pg gateway commit.

* Create egov-pg-service

* Delete egov-pg-service

* Create txt

* added pom.xml for egov-pg-service

* first commit for egov-pg-service.

* added missing file.

* added config for egov-pg-service.

* added db image.

* IFIX-761-Included content type property to the file store upload api

* IFIX-761

* IFIX-735 (#75)

and also nic sms implementation fix
# Conflicts:
#	core-services/egov-notification-sms/src/main/java/org/egov/web/notification/sms/service/impl/NICSMSServiceImpl.java

* IFIX-765

* IFIX-765

* IFIX-766

* IFIX-766

* IFIX-766

* IFIX-769

* IFIX-768

* IFIX-768

* IFIX-71,72: Code changes related to dashboard metrix.

* IFIX-71: Code changes related to dashboard metrix.

* IFIX-779

* IFIX-71: Code changes related to dashboard metrix.

* IFIX-779

* Updated Flutter Auto Upgrade to stable Version

* IFIX-72: Code changes related to dashboard metrix.

* IFIX-784

* IFIX-771

* IFIX-786

* Log4J Vulnerability fix (#122)

* query fixes for notification today colleciton scheluder.

* Today Collection Scheduler query fixes for notification. (#123)

* IFIX-760, 780, 785: code changes related to expense and revenue dashboards.

* name search implementaion.

* null data issu fixed.

* ifix-791 HH name search functionality (#129)

* Rollout dashboard (#119)

* Create txt

* first commit for rollout dashboard

* Update requirements.txt

* Update app.properties

* Update app.properties

* Create readme.md

* Create localsetup.md

* Update and rename readme.md to README.md

* Update and rename localsetup.md to LOCALSETUP.md

* Update LOCALSETUP.md

* Update LOCALSETUP.md

* Update LOCALSETUP.md

* Update README.md

* added config files.

* Update cronjob.yaml

* fix for ifix-777

* IFIX-647

* IFIX-796

* IFIX-485
Rollout dashboard as cron job changes to load the env file and read from .env file

* Ifix 485 (#137)

* IFIX-485
Rollout dashboard as cron job changes to load the env file and read from .env file

* IFIX-485
Rollout dashboard as cron job changes to load the env file and read from .env file

* IFIX-485
Rollout dashboard as cron job changes to load the env file and read from .env file

* Update Dockerfile

* updated docker file

* udpating docker file

udpating docker file

* IFIX-773,781: Code changes Related to Demand generation.

* IFIX-485
Rollout dashboard as cron job changes to load the env file and read from .env file

* IFIX-281: Code changes Related to challan plainsearch.

* IFIX-281: Code changes Related to challan plainsearch.

* Billing service changes (#139)

* Update ApplicationProperties.java

* Update BillServicev2.java

* Update application.properties

* IFIX-281: Code changes Related to challan plainsearch.

* IFIX-694-Added test script files

* IFIX-694

* IFIX-694

* IFIX-686

* IFIX-802: Code changes related to cancel all the demands if payment is not done.

* IFIX-805

* Added update demand property

* IFIX-797

* IFIX-647

* IFIX-801

* IFIX-806

* IFIX-804

* IFIX-811

* IFIX-811

* IFIX-807

* IFIX-806

* IFIX-810

* IFIX-809

* IFIX-813

* IFIX-816

* IFIX-802: Code changes for non-metered connection.

* IFIX-814: Data changes for plain search.

* IFIX-818

* IFIX-819

* IFIX-819

* IFIX-820

* Update SearchCriteria.java

* ifix 817 added new metrix for rollout dashboard. (#167)

* ifix 817 added new metrix for rollout dashboard.

* remove access token properties and method.

* added utc time to localtime

* time zode added for ist

* IFIX-819

* Ifix 823 824 (#169)

* IFIX-832
IFIX-824

Vendor search should return only active vendors
If an employee exists with Vendor Mobile Number then throw the error.

* IFIX-832
IFIX-824

Vendor search should return only active vendors
If an employee exists with Vendor Mobile Number then throw the error.

* Revert "IFIX-832"

This reverts commit fdddc03ad7254416bfcc7fecc1418c8c43d04f37.

* time conversion.

* Ifix 823 824 (#171)

* IFIX-832
IFIX-824

Vendor search should return only active vendors
If an employee exists with Vendor Mobile Number then throw the error.

* IFIX-832
IFIX-824

Vendor search should return only active vendors
If an employee exists with Vendor Mobile Number then throw the error.

* Revert "IFIX-832"

This reverts commit fdddc03ad7254416bfcc7fecc1418c8c43d04f37.

* Query issue fixed

* IFIX-817 (#172)

plain search echallan issue with statelevel tenant

* IFIX-817
plain search echallan issue with statelevel tenant

* IFIX-817 (#172)

plain search echallan issue with statelevel tenant

* IFIX-821

* IFIX-825

* IFIX-824

* IFIX-825

* added pytz script

* IFIX-826

* IFIX-827

* IFIX-830

* IFIX-831

* IFIX-829

* IFIX-829

* IFIX-829

* IFIX-815

* IFIX-815,837:: Code changes for pending collection in table.

* IFIX-838

* IFIX-838

* IFIX-839

* IFIX-845

* IFIX-845

* IFIX-845

* Update Dockerfile

+++ frontend/mgramseva/lib/model/bill/meter_demand_details.dart
@@ -22,11 +22,11 @@ class MeterReadings {
   @JsonKey(name: "meterStatus")
   String? meterStatus;
   @JsonKey(name: "lastReading")
-  double? lastReading;
+  int? lastReading;
   @JsonKey(name: "lastReadingDate")
   int? lastReadingDate;
   @JsonKey(name: "currentReading")
-  double? currentReading;
+  int? currentReading;
   @JsonKey(name: "currentReadingDate")
   int? currentReadingDate;
   @JsonKey(name: "connectionNo")
